feat: added flow provider to wallets
