--- packages/wallets/src/components/Flow/index.ts
@@ -1,2 +1,0 @@
-export { default as WalletFlowSwitcher } from './WalletFlowSwitcher';
-export { default as WalletFlowProvider } from './WalletFlowProvider';
