+++ packages/wallets/src/components/FlowProvider/index.ts
@@ -1,0 +1,4 @@
+import FlowProvider, { useFlow, TFlowProviderContext } from './FlowProvider';
+
+export { FlowProvider, useFlow };
+export type { TFlowProviderContext };
